Rename package to include score name
