+++ core-contracts/DAOfund/build.gradle
@@ -36,7 +36,7 @@ dependencies {
 }
 
 optimizedJar {
-    mainClassName = 'network.balanced.score.core.DAOfund'
+    mainClassName = 'network.balanced.score.core.daofund.DAOfund'
     duplicatesStrategy = DuplicatesStrategy.EXCLUDE
     from {
         configurations.runtimeClasspath.collect{it.isDirectory() ? it : zipTree(it)}
