+++ core-contracts/DAOfund/src/main/java/network/balanced/score/core/daofund/utils/EnumerableSetDB.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package network.balanced.score.core.utils;
+package network.balanced.score.core.daofund.utils;
 
 import score.ArrayDB;
 import score.Context;
